layout/base/tests/test_bug471126.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=471126
-->
<head>
  <title>Test for Bug 471126</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=471126">Mozilla Bug 471126</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 471126 **/

function test()
{
  var selection = window.getSelection();
  selection.collapse(document.documentElement, 0);
  document.documentElement.addEventListener("click", function(){ var foo = window; }, false);
}
test();
ok(true, "Shouldn't leak");

</script>
</pre>
</body>
</html>
